+++ services/horizon/internal/actions/liquidity_pool.go
@@ -68,7 +68,7 @@ type LiquidityPoolsQuery struct {
 
 // URITemplate returns a rfc6570 URI template the query struct
 func (q LiquidityPoolsQuery) URITemplate() string {
-	return "/liquidity_pools?{?reserves,?account}"
+	return "/liquidity_pools?{?reserves,account}"
 }
 
 // Validate validates and parses the query
@@ -122,7 +122,7 @@ func (handler GetLiquidityPoolsHandler) GetResourcePage(w HeaderWriter, r *http.
 
 	query := history.LiquidityPoolsQuery{
 		PageQuery: pq,
-		AccountID: qp.Account,
+		Account:   qp.Account,
 		Assets:    qp.reserves,
 	}